# apps/backend/internal/adapter/httpapi/generated/openapi.codegen.yaml
openapi: 3.0.3
info:
  title: Telomoyo POS API
  version: 1.0.0
  description: |
    The contract used by the Android terminal and the Go backend. Money is integer
    Indonesian rupiah, timestamps are RFC 3339, reporting uses Asia/Jakarta, and
    transaction IDs are raw ULIDs. User-facing surfaces add the `TRX-` prefix.

    Unless a response is a downloadable export, successful JSON responses use
    `{data, meta}` and errors use `{error: {code, message, details, requestId}}`.
servers:
  - url: /api/v1
    description: Version 1 API
security:
  - bearerAuth: []
...
paths:
  /auth/contexts:
    get:
      operationId: getAuthContexts
      tags: [Authentication]
      summary: List active business units and organization-management capabilities
      responses:
        "200":
          description: All active tenants are available to every active account; roles are organization-wide.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableContextsEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /auth/switch-context:
    post:
      operationId: switchAuthContext
      tags: [Authentication]
      summary: Exchange the session into account or selected tenant context
      description: Tenant context always starts in Production. Requires online outbox drainage or explicit quarantine after revoked access. The previous session is revoked and retained as immutable origin evidence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchContextRequest"
      responses:
        "200":
          description: The previous session remains immutable origin evidence. No new platform contexts are issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /auth/upgrade-session:
    post:
      operationId: upgradeAuthSession
      tags: [Authentication]
      summary: Upgrade the session protocol while preserving its selected context
      description: |
        Exchange the session for protocol 3 after the client drains its current
        outbox and finishes printing. Preserve account, tenant, enrollment, mode,
        and active generation; issue transaction_total Sandbox QRIS policy.
        Retain the revoked previous session as immutable signed origin evidence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeSessionRequest"
      responses:
        "200":
          description: Upgraded session with the original context and negotiated payment policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /management/tenants:
    get:
      operationId: listManagedTenants
      tags: [Management]
      summary: List all business units including suspended and pending tenants
      x-required-role: superadmin
      responses:
        "200":
          description: Organization-wide tenant list for an authenticated Superadmin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createManagedTenant
      tags: [Management]
      summary: Create an active business unit with an empty catalog
      description: Requires an authenticated Superadmin and enabled provisioning. Initializes the receipt identity with the management name. No invitation or membership approval is created.
      x-required-role: superadmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Active tenant accessible to all active staff.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /management/tenants/{tenantId}:
    patch:
      operationId: updateManagedTenant
      tags: [Management]
      summary: Change a tenant's management name without changing receipt identity
      description: Authenticated Superadmin only. Requires the current management revision. Tenant ID and slug are immutable; tenant deletion is not supported.
      x-required-role: superadmin
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateManagedTenantRequest"
      responses:
        "200":
          description: Tenant with incremented management revision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /management/tenants/{tenantId}/status:
    post:
      operationId: setManagedTenantStatus
      tags: [Management]
      summary: Activate, suspend, or reactivate a business unit
      description: Authenticated Superadmin only. Activation accepts an existing pending_setup tenant without an invitation. Suspension preserves data and queued evidence; reactivation never revives revoked sessions.
      x-required-role: superadmin
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetTenantStatusRequest"
      responses:
        "200":
          description: Updated tenant status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /management/users:
    get:
      operationId: listManagedUsers
      tags: [Management]
      summary: List organization-wide staff accounts
      x-required-role: superadmin
      responses:
        "200":
          description: Global accounts and roles for an authenticated Superadmin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUserListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createManagedUser
      tags: [Management]
      summary: Create a staff account using a temporary password
      description: Authenticated Superadmin only. The staff account must change its temporary password before operating and can subsequently select every active tenant.
      x-required-role: superadmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          $ref: "#/components/responses/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /management/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      operationId: getManagedUser
      tags: [Management]
      summary: Get a global staff account
      x-required-role: superadmin
      responses:
        "200":
          $ref: "#/components/responses/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateManagedUser
      tags: [Management]
      summary: Change an account's global role or active status
      description: Authenticated Superadmin only. Role changes and deactivation revoke all account sessions. Self-demotion, self-deactivation, and removing the last active Superadmin are prohibited. Account deletion is not supported.
      x-required-role: superadmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateManagedUserRequest"
      responses:
        "200":
          $ref: "#/components/responses/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /management/users/{userId}/reset-password:
    post:
      operationId: resetManagedUserPassword
      tags: [Management]
      summary: Set a temporary password and revoke all of the account's sessions
      description: Authenticated Superadmin only. The account must change this password before operating again. Credentials are never included in audit records.
      x-required-role: superadmin
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          $ref: "#/components/responses/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /management/audit:
    get:
      operationId: listManagementAudit
      tags: [Management]
      summary: List durable organization-wide administrative audit history
      description: Authenticated Superadmin only. Does not expose passwords, invitation codes, or raw QRIS payloads.
      x-required-role: superadmin
      parameters:
        - $ref: "#/components/parameters/AuditLimit"
      responses:
        "200":
          description: Recent account and tenant lifecycle events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformAuditListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/register-invitation:
    post:
      operationId: registerInvitation
      deprecated: true
      tags: [Authentication]
      summary: Retired invitation endpoint; all active accounts can enter every active tenant
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInvitationRequest"
      responses:
        "410":
          $ref: "#/components/responses/Gone"
  /auth/invitations/accept:
    post:
      operationId: acceptInvitation
      deprecated: true
      tags: [Authentication]
      summary: Retired invitation endpoint; all active accounts can enter every active tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationRequest"
      responses:
        "410":
          $ref: "#/components/responses/Gone"
  /platform/tenants:
    get:
      operationId: listPlatformTenants
      deprecated: true
      tags: [Platform]
      summary: Retired administration endpoint; upgrade to organization management
      responses:
        "426":
          $ref: "#/components/responses/UpgradeRequired"

    post:
      operationId: createTenant
      deprecated: true
      tags: [Platform]
      summary: Retired administration endpoint; upgrade to organization management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /platform/tenants/{tenantId}/status:
    post:
      operationId: setTenantStatus
      deprecated: true
      tags: [Platform]
      summary: Retired administration endpoint; upgrade to organization management
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetTenantStatusRequest"
      responses:
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /platform/tenants/{tenantId}/invitation:
    post:
      operationId: reissueOwnerInvitation
      deprecated: true
      tags: [Platform]
      summary: Retired invitation endpoint; all active accounts can enter every active tenant
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "410":
          $ref: "#/components/responses/Gone"
  /platform/audit:
    get:
      operationId: listPlatformAudit
      deprecated: true
      tags: [Platform]
      summary: Retired administration endpoint; upgrade to organization management
      parameters:
        - $ref: "#/components/parameters/AuditLimit"
      responses:
        "426":
          $ref: "#/components/responses/UpgradeRequired"
...
  /tenant/members/{userId}:
    patch:
      operationId: updateTenantMember
      deprecated: true
      tags: [Tenants]
      summary: Retired administration endpoint; upgrade to organization management
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantMemberRequest"
      responses:
        "426":
          $ref: "#/components/responses/UpgradeRequired"
...
  /tenant/invitations/{invitationId}:
    delete:
      operationId: revokeTenantInvitation
      deprecated: true
      tags: [Tenants]
      summary: Retired invitation endpoint; all active accounts can enter every active tenant
      parameters:
        - $ref: "#/components/parameters/InvitationId"
      responses:
        "410":
          $ref: "#/components/responses/Gone"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpaqueSessionToken
      description: Opaque 256-bit token. Only its hash is persisted by the backend.
...
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
...
    BadRequest:
      $ref: "#/components/responses/ErrorResponse"
    Unauthorized:
      $ref: "#/components/responses/ErrorResponse"
    Forbidden:
      $ref: "#/components/responses/ErrorResponse"
    NotFound:
      $ref: "#/components/responses/ErrorResponse"
    Conflict:
      $ref: "#/components/responses/ErrorResponse"
...
    UnprocessableEntity:
      $ref: "#/components/responses/ErrorResponse"
...
  schemas:
...
    AvailableContextsEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AvailableContexts"
        meta:
          $ref: "#/components/schemas/Meta"
...
    UUID:
      type: string
      format: uuid
...